fix#27: api 타입 수정, 사용하지 않는 타입 필드 제거
