+++ src/apis/api.type.ts
@@ -1,0 +1,17 @@
+export interface ApiResponse<T> {
+  success: boolean;
+  message: string;
+  data: T;
+  timestamp: string;
+}
+
+export interface Pagination {
+  currentPage: number;
+  totalPages: number;
+  pageSize: number;
+  totalCount: number;
+  hasMore: boolean;
+  hasPrevious: boolean;
+  first: boolean;
+  last: boolean;
+}
