+++ src/apis/main/main.ts
@@ -1,14 +1,7 @@
 import instance from "../instance";
-import type { DiaryBoardDetail } from "@/apis/community/community.type";
-import type { NewsItem } from "../news/news.type";
-import type { MainPopularCommunityPost } from "./main.type";
+import type { MainPageResponse } from "./main.type";
 
-export interface MainPageResponse {
-  recentDiaries: DiaryBoardDetail[];
-  recentNews: NewsItem[];
-  popularCommunityPosts: MainPopularCommunityPost[];
-}
-
+// 메인 페이지: 실천일지, 커뮤니티 인기글, 뉴스 아이템 조회
 export const getMainPage = async () => {
   const { data } = await instance.get<MainPageResponse>("/api/v1/main");
   return data;
