+++ src/apis/vote/vote.ts
@@ -1,9 +1,8 @@
 import instance from "@/apis/instance";
 import type {
   OdaVoteResponse,
-  OdaVoteRequest,
   DiaryVoteResponse,
-  DiaryVoteRequest,
+  VoteRequest,
   MonthlyVoteResultResponse,
 } from "./vote.type";
 
@@ -15,27 +14,25 @@ export const fetchOdaVote = () => {
 };
 
 // ODA 투표 참여
-export const voteOda = (body: OdaVoteRequest) => {
+export const voteOda = (body: VoteRequest) => {
   return instance.post("/api/v1/oda-vote/vote", body);
 };
 
 // 실천일지 투표 조회
 export const fetchDiaryVote = () => {
   return instance.get<{
-    success: boolean;
-    message: string;
     data: DiaryVoteResponse;
   }>("/api/v1/monthly-vote/current");
 };
 
 // 실천일지 투표 참여
-export const voteDiary = (body: DiaryVoteRequest) => {
+export const voteDiary = (body: VoteRequest) => {
   return instance.post("/api/v1/monthly-vote/vote", body);
 };
 
 //월별 투표 결과 조회
 export const fetchMonthlyVoteResult = (year: number, month: number) => {
-  return instance.get<{ success: boolean; data: MonthlyVoteResultResponse }>(
+  return instance.get<{ data: MonthlyVoteResultResponse }>(
     `/api/v1/monthly-vote/result/${year}/${month}`
   );
 };
