+++ src/apis/vote/vote.type.ts
@@ -1,105 +1,52 @@
+// Oda 투표 후보 리스트 조회
+export interface OdaVoteResponse {
+  id: number;
+  candidates: OdaVoteCandidate[];
+  hasUserVoted: boolean;
+  userVotedCandidateId: number | null;
+}
+
+// Oda 투표 후보 타입
 export interface OdaVoteCandidate {
   id: number;
   odaProject: {
     id: number;
     title: string;
-    content: string;
     url: string;
-    category: string;
-    countryName: string;
-    projectStartDate: string;
-    projectEndDate: string;
-    budget: string;
-    publishDate: string;
-    matchScore: number;
     summary: string;
   };
-  voteCount: number;
-  votePercentage: number;
-  rank: number;
 }
 
-export interface OdaVoteResponse {
+// 실천일지 투표 후보 리스트 조회
+export interface DiaryVoteResponse {
   id: number;
-  year: number;
-  month: number;
   title: string;
-  description: string;
-  status: string;
-  startDate: string;
-  endDate: string;
-  createdAt: string;
-  candidates: OdaVoteCandidate[];
-  totalVoteCount: number;
+  candidates: DiaryVoteCandidate[];
   hasUserVoted: boolean;
   userVotedCandidateId: number | null;
 }
 
-export interface OdaVoteRequest {
-  candidateId: number;
-}
-
+// 실천일지 후보 타입
 export interface DiaryVoteCandidate {
   candidateId: number;
   diaryId: number;
   diaryTitle: string;
   diaryDescription: string;
-  diaryAction: string;
   authorName: string;
-  diaryCreatedAt: string;
-  diaryLikes: number;
-  diaryViewCount: number;
-  voteCount: number;
-  ranking: number;
 }
 
-export interface DiaryVoteResponse {
-  id: number;
-  year: number;
-  month: number;
-  title: string;
-  description: string;
-  status: string;
-  startDate: string;
-  endDate: string;
-  totalVotes: number;
-  candidates: DiaryVoteCandidate[];
-  hasUserVoted: boolean;
-  userVotedCandidateId: number | null;
-  userVotedAt: string;
-}
-
-export interface DiaryVoteRequest {
+// 투표하기 타입
+export interface VoteRequest {
   candidateId: number;
 }
 
-
+// 월별 투표 결과 조회 타입
 export interface MonthlyVoteResultResponse {
   id: number;
-  year: number;
-  month: number;
+  // 이달의 외교 실천왕
   title: string;
-  description: string;
-  status: string;
-  startDate: string;
-  endDate: string;
-  totalVotes: number;
   candidates: DiaryVoteCandidate[];
-
-  odaVoteId: number;
+  // 의미있는 ODA 사업 투표
   odaVoteTitle: string;
-  odaVoteDescription: string;
-  odaVoteStatus: string;
-  odaVoteStartDate: string;
-  odaVoteEndDate: string;
-  odaTotalVotes: number;
   odaCandidates: OdaVoteCandidate[];
-
-  hasUserVoted: boolean;
-  userVotedCandidateId: number | null;
-  userVotedAt: string | null;
-
-  hasUserVotedOda: boolean;
-  userVotedOdaCandidateId: number | null;
-  userVotedOdaAt: string | null;
 }
