+++ src/pages/test/Result.tsx
@@ -127,7 +127,6 @@ export default function TestResult() {
                     id: 0,
                     newsId: 0,
                     scrapId: 0,
-                    scrapedAt: "",
                     category: "",
                     categoryDisplay: "",
                     title: news.title,
@@ -153,7 +152,6 @@ export default function TestResult() {
                     id: 0,
                     newsId: 0,
                     scrapId: 0,
-                    scrapedAt: "",
                     category: "",
                     categoryDisplay: "",
                     title: prog.businessName,
